Switch to LlamaIndex Retriever
